+++ src/Spago/Purs.hs
@@ -82,6 +82,7 @@ data DocsFormat
   | Markdown
   | Ctags
   | Etags
+  deriving (Eq)
 
 parseDocsFormat :: Text -> Maybe DocsFormat
 parseDocsFormat = \case
@@ -92,7 +93,7 @@ parseDocsFormat = \case
   _          -> Nothing
 
 
-docs :: Spago m => Maybe DocsFormat -> [SourcePath] -> m ()
+docs :: Spago m => DocsFormat -> [SourcePath] -> m ()
 docs format sourcePaths = do
   let
     printDocsFormat :: DocsFormat -> Text
@@ -103,7 +104,7 @@ docs format sourcePaths = do
       Etags    -> "etags"
 
     paths = Text.intercalate " " $ surroundQuote <$> map unSourcePath sourcePaths
-    formatStr = printDocsFormat $ fromMaybe Html format
+    formatStr = printDocsFormat format
     cmd = "purs docs " <> paths <> " --format " <> formatStr
   runWithOutput cmd
     "Docs generation succeeded."